+++ packages/engine-multi/src/worker/worker.ts
@@ -21,6 +21,7 @@ type RunOptions = {
   adaptorPaths: Record<string, { path: string }>;
   whitelist?: RegExp[];
   sanitize: SanitizePolicies;
+  statePropsToRemove?: string[];
   // TODO timeout
 };
 
@@ -29,7 +30,8 @@ workerpool.worker({
   handshake: () => true,
 
   run: (plan: ExecutionPlan, runOptions: RunOptions) => {
-    const { adaptorPaths, whitelist, sanitize } = runOptions;
+    const { adaptorPaths, whitelist, sanitize, statePropsToRemove } =
+      runOptions;
     const { logger, jobLogger } = createLoggers(plan.id!, sanitize);
     const options = {
       strict: false,
@@ -40,6 +42,7 @@ workerpool.worker({
         whitelist,
         cacheKey: plan.id,
       },
+      statePropsToRemove,
       callbacks: {
         // TODO: this won't actually work across the worker boundary
         // For now I am preloading credentials