+++ packages/ws-worker/CHANGELOG.md
@@ -1,5 +1,14 @@
 # ws-worker
 
+## 0.5.0
+
+### Minor Changes
+
+- Add state-props-to-remove option
+- Updated dependencies [56b6e44]
+  - @openfn/engine-multi@0.2.6
+  - @openfn/runtime@0.2.4
+
 ## 0.4.0
 
 ### Minor Changes